+++ src/index.ts
@@ -2,6 +2,7 @@ export { createApp } from './app';
 export type { App, Plugin } from './app';
 export { store } from './store';
 export { nextTick } from './scheduler';
+export { watchEffect } from './effect';
 export { reactive } from '@vue/reactivity';
 export { devtools, disableDevtools } from './devtools';
 export type { LiteVueDevtools, DevtoolsEvent } from './devtools';